frontend/requirements.txt
absl-py==1.0.0
appengine-python-standard==1.0.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cloudstorage==0.11.0
colorama==0.4.4
cycler==0.11.0
Cython==0.29.30
debugpy==1.6.2
decorator==5.1.1
distlib==0.3.4
entrypoints==0.4
executing==0.8.3
filelock==3.7.1
Flask==2.1.2
flatbuffers==2.0
fonttools==4.34.4
frozendict==2.3.2
gast==0.5.3
google-api-core==2.7.1
google-api-python-client==2.41.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-datastore==2.8.0
google-cloud-storage==2.4.0
google-cloud-vision==2.7.3
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.55.0
grpcio==1.47.0
grpcio-status==1.47.0
h5py==3.6.0
httplib2==0.20.4
idna==3.3
inflection==0.5.1
install==1.3.5
ipykernel==6.15.1
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.11.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mediapipe==0.8.10.1
mock==4.0.3
nest-asyncio==1.5.5
nltk==3.7
numpy==1.22.2
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
proto-plus==1.20.6
protobuf==3.19.4
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.7
pyTelegramBotAPI==4.4.0
pytesseract @ git+https://github.com/madmaze/pytesseract.git@799bf03a0f5a14cb504889cd00608c187a9444e5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-magic==0.4.27
pytz==2021.3
pywin32==304
pyzmq==23.2.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
stack-data==0.3.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.63.0
traitlets==5.3.0
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.15.1
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.13.3